\name{IVaids}
\alias{IVaids}
\title{Data: AIDS patient among IV drug user}
\description{
        232 patients infected with HIV. Time to AIDS.
136 left AIDS-free. 14 died with AIDS. 82 had AIDS while in the program.
}
\usage{ data(IVaids) }
\references{ 
Julia, Olga and Gomez, Guadalupe (2008)
}
\keyword{datasets}
