\name{IVaids}
\alias{IVaids}
\title{Data: AIDS patient among IV drug user}
\description{
     computes the self-consistent estimator of CDF from doubly censored data
         via an algorithm specified in LUPE 2008.
         It only have jumps on the observed failure times (where d=1).

  When there are ties in the observed times, the left (right)
 censored points are treated as happened before (after) the failure time, 
 to break tie.
}
\usage{ data(IVaids) }
\references{ 
Lupe (2008)
}
\keyword{datasets}
