\name{IVaids}
\alias{IVaids}
\title{Data: AIDS patient among IV drug user}
\description{
  Time to AIDS among 232 patients infected with HIV. 
136 left AIDS-free. 14 died with AIDS. 82 had AIDS while in the program.
}
\usage{ data(IVaids) }
\references{ 
Julia, Olga and Gomez, Guadalupe (2011)
Simultaneous marginal survival estimators when doubly censored data is present.
Lifetime Data Analysis,
July 2011, Volume 17, Issue 3, pp 347-372.
}
\keyword{datasets}
